+++ R/get_wcde_single.R
@@ -5,11 +5,13 @@
 #' @param scenario Vector with a numbers corresponding the scenarios. See details in `wcde` for more information.
 #' @param country_code Vector of length one or more of country numeric codes based on ISO 3 digit numeric values.
 #' @param server Character string for server to download from. Defaults to `iiasa`, but can use `github` if IIASA server is down.
+#' @param version Character string for version of projections to obtain. Defaults to `wcde-v2`, but can use `wcde-v2` and `fume`. Scenario and indicator availability vary between versions.
 #'
 #' @return A tibble with multiple columns.
 #' @keywords internal
 #' @export
-get_wcde_single <- function(indicator = NULL, scenario = 2, country_code = NULL, server = NULL){
+get_wcde_single <- function(indicator = NULL, scenario = 2, country_code = NULL,
+                            server = NULL, version = NULL){
   # scenario = c(1, 3); indicator = "tfr"; country_code = c(40, 100)
   # scenario = 2; indicator = "e0"; country_code = "900"
   # server = "github"
@@ -63,15 +65,22 @@ get_wcde_single <- function(indicator = NULL, scenario = 2, country_code = NULL,
   pb <- progress::progress_bar$new(total = nrow(d0))
   pb$tick(0)
   d0 <- d0 %>%
-    dplyr::mutate(u = paste0("http://dataexplorer.wittgensteincentre.org/wcde-data/data-single/",
-                             scenario, "/", indicator, "/", country_code, ".csv")) %>%
+    dplyr::mutate(u = paste0("http://dataexplorer.wittgensteincentre.org/wcde-data/",
+                             version, "/data-single/", scenario, "/",
+                             indicator, "/", country_code, ".csv")) %>%
     {if(server == "github")
-      dplyr::mutate(., u = paste0("https://github.com/guyabel/wcde/raw/main/data-host/",
-                                  scenario, "/", indicator, "/", country_code, ".csv"))
+      dplyr::mutate(., u = paste0("https://github.com/guyabel/wcde-data/raw/main/",
+                                  version, "/data-single/", scenario, "/",
+                                  indicator, "/", country_code, ".csv"))
+      else .} %>%
+    {if(server == "iiasa-local")
+      dplyr::mutate(., u = paste0("../wcde-data/",
+                                  version, "/data-single/", scenario, "/",
+                                  indicator, "/", country_code, ".csv"))
       else .} %>%
     dplyr::mutate(d = purrr::map(.x = u, .f = ~read_with_progress(f = .x))) %>%
     dplyr::group_by(scenario) %>%
-    dplyr::reframe(dplyr::bind_cols(d), .groups = "drop_last") %>%
+    dplyr::reframe(dplyr::bind_cols(d)) %>%
     dplyr::ungroup()
   pb$terminate()