+++ NEWS.md
@@ -31,3 +31,4 @@
 * Add `server` argument in `get_wcde()`
 * Updated UK pop-age file for scenario 1 on server
 * Replaced `dplyr::summarize()` with `dplyr::reframe()` in `get_wcde_single()`
+* Add `version` argument in `get_wcde()`